prototypes/frank/requirements.txt
flask==3.1.0
sqlalchemy==2.0.36
httpx==0.28.1
python-dotenv==1.0.1
